requirements.txt
ga4gh.vrs[extras]==2.0.0a3
pyVCF3
diskcache
biocommons.seqrepo
glom
click